fix(dsp): cast WOLA synthesis window to buffer dtype (#4107)
In the psychoacoustic-EQ WOLA path the overlap-add buffer is created with
dtype=audio.dtype, but the synthesis window (hann(chunk_size)) was float64.
Multiplying the processed chunk by a float64 window promotes the product to
float64, which the in-place += into the float32 buffer then truncates on every
overlap step (4 contributions per sample at 75% overlap) — avoidable
quantization noise and a float64 leak into a float32 path.

Cast the window to audio.dtype (astype(..., copy=False)) so the accumulation
stays in the input dtype.

Sibling check: the other hann/hanning windows (spectral.py, resonance_notcher,
psychoacoustic_eq analysis window, interpolation_helpers) are analysis-only
(FFT magnitude/PSD) or a single control-envelope multiply — none do per-overlap
output accumulation, so float64 there is intended.

Tests: added float32 and float64 WOLA dtype-preservation cases asserting the
chunk handed to the EQ and the output keep the input dtype and sample count.

Closes #4107

+++ auralis/core/processing/eq_processor.py
@@ -189,7 +189,11 @@ class EQProcessor:
         # output is trimmed to original_length (#3437).
         out_shape = (original_length + chunk_size,) + audio.shape[1:]
         processed_audio = np.zeros(out_shape, dtype=audio.dtype)
-        wola_window = hann(chunk_size)
+        # Cast the synthesis window to the buffer dtype so the overlap-add
+        # multiply stays in float32: a float64 window promotes each contribution
+        # to float64, which the in-place += into the float32 buffer then
+        # truncates on every overlap step (#4107).
+        wola_window = hann(chunk_size).astype(audio.dtype, copy=False)
 
         for i in range(0, original_length, hop_size):
             end_idx = min(i + chunk_size, original_length)
